+++ weblate/trans/views/reports.py
@@ -30,7 +30,7 @@ def format_plaintext_join(sep, format_string, args_generator):
     return sep.join(format_plaintext(format_string, *args) for args in args_generator)
 
 
-def generate_credits(user, start_date, end_date, **kwargs):
+def generate_credits(user, start_date, end_date, language_code: str, **kwargs):
     """Generate credits data for given component."""
     result = []
 
@@ -38,7 +38,11 @@ def generate_credits(user, start_date, end_date, **kwargs):
     if user:
         base = base.filter(author=user)
 
-    for language in Language.objects.filter(**kwargs).distinct().iterator():
+    languages = Language.objects.filter(**kwargs)
+    if language_code:
+        languages = languages.filter(code=language_code)
+
+    for language in languages.distinct().iterator():
         authors = base.filter(language=language, **kwargs).authors_list(
             (start_date, end_date)
         )
@@ -56,14 +60,17 @@ def get_credits(request, project=None, component=None):
     if project is None:
         obj = None
         kwargs = {"translation__isnull": False}
+        scope = {}
     elif component is None:
         obj = get_project(request, project)
         kwargs = {"translation__component__project": obj}
+        scope = {"project": obj}
     else:
         obj = get_component(request, project, component)
         kwargs = {"translation__component": obj}
+        scope = {"component": obj}
 
-    form = ReportsForm(request.POST)
+    form = ReportsForm(scope, request.POST)
 
     if not form.is_valid():
         show_form_errors(request, form)
@@ -73,6 +80,7 @@ def get_credits(request, project=None, component=None):
         None if request.user.has_perm("reports.view", obj) else request.user,
         form.cleaned_data["start_date"],
         form.cleaned_data["end_date"],
+        form.cleaned_data["language"],
         **kwargs,
     )
 
@@ -156,13 +164,15 @@ COUNT_DEFAULTS = {
 }
 
 
-def generate_counts(user, start_date, end_date, **kwargs):
+def generate_counts(user, start_date, end_date, language_code: str, **kwargs):
     """Generate credits data for given component."""
     result = {}
     action_map = {Change.ACTION_NEW: "new", Change.ACTION_APPROVE: "approve"}
 
     base = Change.objects.content().filter(unit__isnull=False)
     base = base.filter(author=user) if user else base.filter(author__isnull=False)
+    if language_code:
+        Change.objects.filter(language__code=language_code)
 
     changes = base.filter(
         timestamp__range=(start_date, end_date), **kwargs
@@ -215,7 +225,7 @@ def get_counts(request, project=None, component=None):
         obj = get_component(request, project, component)
         kwargs = {"component": obj}
 
-    form = ReportsForm(request.POST)
+    form = ReportsForm(kwargs, request.POST)
 
     if not form.is_valid():
         show_form_errors(request, form)
@@ -225,6 +235,7 @@ def get_counts(request, project=None, component=None):
         None if request.user.has_perm("reports.view", obj) else request.user,
         form.cleaned_data["start_date"],
         form.cleaned_data["end_date"],
+        form.cleaned_data["language"],
         **kwargs,
     )